refactor: decompose component with settings
