+++ client/src/components/ui/value-card/value-cardData.ts
@@ -3,27 +3,27 @@ import { IValueCard } from '@models/types';
 export const defaultValueForCustomDeck = ['PASS', '?', '1', '10', '50'];
 export const valueCardData: IValueCard[] = [
   {
-    valueCardID: 1,
+    valueCardID: '1',
     name: 'SP',
     value: 'PASS',
   },
   {
-    valueCardID: 2,
+    valueCardID: '2',
     name: 'SP',
     value: '?',
   },
   {
-    valueCardID: 3,
+    valueCardID: '3',
     name: 'SP',
     value: 1,
   },
   {
-    valueCardID: 4,
+    valueCardID: '4',
     name: 'SP',
     value: 10,
   },
   {
-    valueCardID: 5,
+    valueCardID: '5',
     name: 'SP',
     value: 50,
   },
